Fix standard gate tests not running for all simulation methods
Standard gate tests were not being passed the test class backend options in their exeuction method. This means that they were always running on the automatic simulation method.
