+++ test/terra/backends/qasm_simulator/qasm_algorithms.py
@@ -33,7 +33,8 @@ class QasmAlgorithmTests:
         circuits = ref_algorithms.grovers_circuit(
             final_measure=True, allow_sampling=True)
         targets = ref_algorithms.grovers_counts(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -43,7 +44,8 @@ class QasmAlgorithmTests:
         shots = 4000
         circuits = ref_algorithms.teleport_circuit()
         targets = ref_algorithms.teleport_counts(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -69,7 +71,8 @@ class QasmAlgorithmTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -83,7 +86,8 @@ class QasmAlgorithmTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -105,7 +109,8 @@ class QasmAlgorithmTestsMinimalBasis:
             final_measure=True, allow_sampling=True)
         targets = ref_algorithms.grovers_counts(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -116,7 +121,8 @@ class QasmAlgorithmTestsMinimalBasis:
         circuits = ref_algorithms.teleport_circuit()
         targets = ref_algorithms.teleport_counts(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
