+++ test/terra/backends/qasm_simulator/qasm_cliffords.py
@@ -34,7 +34,8 @@ class QasmCliffordTests:
         circuits = ref_1q_clifford.h_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_1q_clifford.h_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -45,7 +46,8 @@ class QasmCliffordTests:
         circuits = ref_1q_clifford.h_gate_circuits_nondeterministic(
             final_measure=True)
         targets = ref_1q_clifford.h_gate_counts_nondeterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -73,7 +75,8 @@ class QasmCliffordTests:
         circuits = ref_1q_clifford.z_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_1q_clifford.z_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -87,7 +90,8 @@ class QasmCliffordTests:
         circuits = ref_1q_clifford.y_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_1q_clifford.y_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -101,7 +105,8 @@ class QasmCliffordTests:
         circuits = ref_1q_clifford.s_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_1q_clifford.s_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -116,7 +121,7 @@ class QasmCliffordTests:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -134,7 +139,7 @@ class QasmCliffordTests:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -145,7 +150,8 @@ class QasmCliffordTests:
         circuits = ref_1q_clifford.sdg_gate_circuits_nondeterministic(
             final_measure=True)
         targets = ref_1q_clifford.sdg_gate_counts_nondeterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -159,7 +165,8 @@ class QasmCliffordTests:
         circuits = ref_2q_clifford.cx_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_2q_clifford.cx_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -170,7 +177,8 @@ class QasmCliffordTests:
         circuits = ref_2q_clifford.cx_gate_circuits_nondeterministic(
             final_measure=True)
         targets = ref_2q_clifford.cx_gate_counts_nondeterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -184,7 +192,8 @@ class QasmCliffordTests:
         circuits = ref_2q_clifford.cz_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_2q_clifford.cz_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -195,7 +204,8 @@ class QasmCliffordTests:
         circuits = ref_2q_clifford.cz_gate_circuits_nondeterministic(
             final_measure=True)
         targets = ref_2q_clifford.cz_gate_counts_nondeterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -209,7 +219,8 @@ class QasmCliffordTests:
         circuits = ref_2q_clifford.swap_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_2q_clifford.swap_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -220,7 +231,8 @@ class QasmCliffordTests:
         circuits = ref_2q_clifford.swap_gate_circuits_nondeterministic(
             final_measure=True)
         targets = ref_2q_clifford.swap_gate_counts_nondeterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -245,7 +257,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -260,7 +273,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -278,7 +292,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -297,7 +312,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -309,7 +325,8 @@ class QasmCliffordTestsWaltzBasis:
             final_measure=True)
         targets = ref_1q_clifford.z_gate_counts_deterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -323,7 +340,8 @@ class QasmCliffordTestsWaltzBasis:
         circuits = ref_1q_clifford.y_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_1q_clifford.y_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -338,7 +356,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -356,7 +375,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -371,7 +391,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -404,7 +425,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -422,7 +444,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -470,7 +493,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -488,7 +512,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -503,7 +528,8 @@ class QasmCliffordTestsWaltzBasis:
             circuits,
             self.SIMULATOR,
             shots=shots,
-            basis_gates=['u1', 'u2', 'u3', 'cx'])
+            basis_gates=['u1', 'u2', 'u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -526,7 +552,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_1q_clifford.h_gate_counts_deterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -538,7 +565,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_1q_clifford.h_gate_counts_nondeterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -569,7 +597,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_1q_clifford.z_gate_counts_deterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -585,7 +614,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_1q_clifford.y_gate_counts_deterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -601,7 +631,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_1q_clifford.s_gate_counts_deterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -613,7 +644,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_1q_clifford.s_gate_counts_nondeterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -628,7 +660,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_1q_clifford.sdg_gate_counts_deterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -640,7 +673,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_1q_clifford.sdg_gate_counts_nondeterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -655,7 +689,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_2q_clifford.cx_gate_counts_deterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -667,7 +702,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_2q_clifford.cx_gate_counts_nondeterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -682,7 +718,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_2q_clifford.cz_gate_counts_deterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -694,7 +731,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_2q_clifford.cz_gate_counts_nondeterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -709,7 +747,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_2q_clifford.swap_gate_counts_deterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -721,32 +760,8 @@ class QasmCliffordTestsMinimalBasis:
             final_measure=True)
         targets = ref_2q_clifford.swap_gate_counts_nondeterministic(shots)
         job = execute(
-            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'])
-        result = job.result()
-        self.assertTrue(getattr(result, 'success', False))
-        self.compare_counts(result, circuits, targets, delta=0.05 * shots)
-
-    # ---------------------------------------------------------------------
-    # Test multiplexer-cx-gate
-    # ---------------------------------------------------------------------
-    def test_multiplexer_cx_gate_deterministic_default_basis_gates(self):
-        """Test cx-gate circuits compiling to backend default basis_gates."""
-        shots = 100
-        circuits = ref_2q_clifford.multiplexer_cx_gate_circuits_deterministic(
-            final_measure=True)
-        targets = ref_2q_clifford.multiplexer_cx_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
-        result = job.result()
-        self.assertTrue(getattr(result, 'success', False))
-        self.compare_counts(result, circuits, targets, delta=0)
-
-    def test_multiplexer_cx_gate_nondeterministic_default_basis_gates(self):
-        """Test cx-gate circuits compiling to backend default basis_gates."""
-        shots = 4000
-        circuits = ref_2q_clifford.multiplexer_cx_gate_circuits_nondeterministic(
-            final_measure=True)
-        targets = ref_2q_clifford.multiplexer_cx_gate_counts_nondeterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+            circuits, self.SIMULATOR, shots=shots, basis_gates=['u3', 'cx'],
+            backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
