+++ test/terra/backends/qasm_simulator/qasm_noncliffords.py
@@ -18,8 +18,8 @@ from qiskit import execute
 from qiskit.providers.aer import QasmSimulator
 
 
-class QasmNonCliffordTests:
-    """QasmSimulator non-Clifford gate tests in default basis."""
+class QasmNonCliffordTestsTGate:
+    """QasmSimulator T gate tests in default basis."""
 
     SIMULATOR = QasmSimulator()
     BACKEND_OPTS = {}
@@ -33,7 +33,8 @@ class QasmNonCliffordTests:
         circuits = ref_non_clifford.t_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_non_clifford.t_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -44,7 +45,8 @@ class QasmNonCliffordTests:
         circuits = ref_non_clifford.t_gate_circuits_nondeterministic(
             final_measure=True)
         targets = ref_non_clifford.t_gate_counts_nondeterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -58,7 +60,8 @@ class QasmNonCliffordTests:
         circuits = ref_non_clifford.tdg_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_non_clifford.tdg_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -69,11 +72,18 @@ class QasmNonCliffordTests:
         circuits = ref_non_clifford.tdg_gate_circuits_nondeterministic(
             final_measure=True)
         targets = ref_non_clifford.tdg_gate_counts_nondeterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
+
+class QasmNonCliffordTestsCCXGate:
+    """QasmSimulator CCX gate tests in default basis."""
+
+    SIMULATOR = QasmSimulator()
+    BACKEND_OPTS = {}
     # ---------------------------------------------------------------------
     # Test ccx-gate
     # ---------------------------------------------------------------------
@@ -83,7 +93,8 @@ class QasmNonCliffordTests:
         circuits = ref_non_clifford.ccx_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_non_clifford.ccx_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -94,11 +105,18 @@ class QasmNonCliffordTests:
         circuits = ref_non_clifford.ccx_gate_circuits_nondeterministic(
             final_measure=True)
         targets = ref_non_clifford.ccx_gate_counts_nondeterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
+
+class QasmNonCliffordTestsCGates:
+    """QasmSimulator CSwap gate tests in default basis."""
+
+    SIMULATOR = QasmSimulator()
+    BACKEND_OPTS = {}
     # ---------------------------------------------------------------------
     # Test cswap-gate (Fredkin)
     # ---------------------------------------------------------------------
@@ -107,7 +125,8 @@ class QasmNonCliffordTests:
         circuits = ref_non_clifford.cswap_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_non_clifford.cswap_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -117,7 +136,8 @@ class QasmNonCliffordTests:
         circuits = ref_non_clifford.cswap_gate_circuits_nondeterministic(
             final_measure=True)
         targets = ref_non_clifford.cswap_gate_counts_nondeterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -131,7 +151,8 @@ class QasmNonCliffordTests:
         circuits = ref_non_clifford.cu1_gate_circuits_nondeterministic(
             final_measure=True)
         targets = ref_non_clifford.cu1_gate_counts_nondeterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -145,7 +166,8 @@ class QasmNonCliffordTests:
         circuits = ref_non_clifford.cu3_gate_circuits_deterministic(
             final_measure=True)
         targets = ref_non_clifford.cu3_gate_counts_deterministic(shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
+        job = execute(circuits, self.SIMULATOR, shots=shots,
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -169,7 +191,8 @@ class QasmNonCliffordTestsWaltzBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u1', 'u2', 'u3', 'cx'])
+                      basis_gates=['u1', 'u2', 'u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -183,7 +206,8 @@ class QasmNonCliffordTestsWaltzBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u1', 'u2', 'u3', 'cx'])
+                      basis_gates=['u1', 'u2', 'u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -200,7 +224,8 @@ class QasmNonCliffordTestsWaltzBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u1', 'u2', 'u3', 'cx'])
+                      basis_gates=['u1', 'u2', 'u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -214,7 +239,8 @@ class QasmNonCliffordTestsWaltzBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u1', 'u2', 'u3', 'cx'])
+                      basis_gates=['u1', 'u2', 'u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -231,7 +257,8 @@ class QasmNonCliffordTestsWaltzBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u1', 'u2', 'u3', 'cx'])
+                      basis_gates=['u1', 'u2', 'u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -245,7 +272,8 @@ class QasmNonCliffordTestsWaltzBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u1', 'u2', 'u3', 'cx'])
+                      basis_gates=['u1', 'u2', 'u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -262,7 +290,8 @@ class QasmNonCliffordTestsWaltzBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u1', 'u2', 'u3', 'cx'])
+                      basis_gates=['u1', 'u2', 'u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -276,7 +305,8 @@ class QasmNonCliffordTestsWaltzBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u1', 'u2', 'u3', 'cx'])
+                      basis_gates=['u1', 'u2', 'u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -293,7 +323,8 @@ class QasmNonCliffordTestsWaltzBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u1', 'u2', 'u3', 'cx'])
+                      basis_gates=['u1', 'u2', 'u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -301,7 +332,7 @@ class QasmNonCliffordTestsWaltzBasis:
     # ---------------------------------------------------------------------
     # Test cu3 gate
     # ---------------------------------------------------------------------
-    def test_cu3_gate_deterministic_default_basis_gates(self):
+    def test_cu3_gate_deterministic_waltz_basis_gates(self):
         """Test cu3-gate gate circuits compiling to u1,u2,u3,cx."""
         shots = 100
         circuits = ref_non_clifford.cu3_gate_circuits_deterministic(
@@ -310,7 +341,8 @@ class QasmNonCliffordTestsWaltzBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u1', 'u2', 'u3', 'cx'])
+                      basis_gates=['u1', 'u2', 'u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -348,7 +380,8 @@ class QasmNonCliffordTestsMinimalBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u3', 'cx'])
+                      basis_gates=['u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -365,7 +398,8 @@ class QasmNonCliffordTestsMinimalBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u3', 'cx'])
+                      basis_gates=['u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -379,7 +413,8 @@ class QasmNonCliffordTestsMinimalBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u3', 'cx'])
+                      basis_gates=['u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
@@ -396,7 +431,8 @@ class QasmNonCliffordTestsMinimalBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u3', 'cx'])
+                      basis_gates=['u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -410,7 +446,8 @@ class QasmNonCliffordTestsMinimalBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u3', 'cx'])
+                      basis_gates=['u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.1 * shots)
@@ -427,37 +464,11 @@ class QasmNonCliffordTestsMinimalBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u3', 'cx'])
+                      basis_gates=['u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.1 * shots)
-
-    # ---------------------------------------------------------------------
-    # Test multiplexer-gate
-    # ---------------------------------------------------------------------
-    def test_multiplexer_cxx_gate_deterministic_default_basis_gates(self):
-        """Test multiplexer-gate gate circuits """
-        shots = 100
-        circuits = ref_non_clifford.multiplexer_ccx_gate_circuits_deterministic(
-            final_measure=True)
-        targets = ref_non_clifford.multiplexer_ccx_gate_counts_deterministic(
-            shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
-        result = job.result()
-        self.assertTrue(getattr(result, 'success', False))
-        self.compare_counts(result, circuits, targets, delta=0)
-
-    def test_multiplexer_cxx_gate_nondeterministic_default_basis_gates(self):
-        """Test ccx-gate gate circuits """
-        shots = 4000
-        circuits = ref_non_clifford.multiplexer_ccx_gate_circuits_nondeterministic(
-            final_measure=True)
-        targets = ref_non_clifford.multiplexer_ccx_gate_counts_nondeterministic(
-            shots)
-        job = execute(circuits, self.SIMULATOR, shots=shots)
-        result = job.result()
-        self.assertTrue(getattr(result, 'success', False))
-        self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
     # ---------------------------------------------------------------------
     # Test cswap-gate (Fredkin)
@@ -471,7 +482,8 @@ class QasmNonCliffordTestsMinimalBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u3', 'cx'])
+                      basis_gates=['u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
@@ -485,7 +497,8 @@ class QasmNonCliffordTestsMinimalBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u3', 'cx'])
+                      basis_gates=['u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.1 * shots)
@@ -502,7 +515,8 @@ class QasmNonCliffordTestsMinimalBasis:
         job = execute(circuits,
                       self.SIMULATOR,
                       shots=shots,
-                      basis_gates=['u3', 'cx'])
+                      basis_gates=['u3', 'cx'],
+                      backend_options=self.BACKEND_OPTS)
         result = job.result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
