+++ test/terra/backends/qasm_simulator/qasm_unitary_gate.py
@@ -21,8 +21,9 @@ from qiskit.providers.aer import QasmSimulator
 
 import numpy as np
 
+
 class QasmUnitaryGateTests:
-    """QasmSimulator additional tests."""
+    """QasmSimulator unitary gate tests."""
 
     SIMULATOR = QasmSimulator()
     BACKEND_OPTS = {}
@@ -38,7 +39,8 @@ class QasmUnitaryGateTests:
             final_measure=True)
         targets = ref_unitary_gate.unitary_gate_counts_deterministic(
             shots)
-        result = execute(circuits, self.SIMULATOR, shots=shots).result()
+        result = execute(circuits, self.SIMULATOR, shots=shots,
+                         backend_options=self.BACKEND_OPTS).result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
 
@@ -47,13 +49,14 @@ class QasmUnitaryGateTests:
         shots = 4000
         circuits = ref_unitary_gate.unitary_random_gate_circuits_nondeterministic(final_measure=True)
         targets = ref_unitary_gate.unitary_random_gate_counts_nondeterministic(shots)
-        result = execute(circuits, self.SIMULATOR, shots=shots).result()
+        result = execute(circuits, self.SIMULATOR, shots=shots,
+                         backend_options=self.BACKEND_OPTS).result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0.05 * shots)
 
 
 class QasmDiagonalGateTests:
-    """QasmSimulator additional tests."""
+    """QasmSimulator diagonal gate tests."""
 
     SIMULATOR = QasmSimulator()
     BACKEND_OPTS = {}
@@ -69,6 +72,7 @@ class QasmDiagonalGateTests:
             final_measure=True)
         targets = ref_diagonal_gate.diagonal_gate_counts_deterministic(
             shots)
-        result = execute(circuits, self.SIMULATOR, shots=shots).result()
+        result = execute(circuits, self.SIMULATOR, shots=shots,
+                         backend_options=self.BACKEND_OPTS).result()
         self.assertTrue(getattr(result, 'success', False))
         self.compare_counts(result, circuits, targets, delta=0)
