+++ test/terra/backends/test_qasm_simulator_stabilizer.py
@@ -48,7 +48,6 @@ class TestQasmStabilizerSimulator(common.QiskitAerTestCase,
                                   QasmResetTests,
                                   QasmConditionalGateTests,
                                   QasmCliffordTests,
-                                  QasmAlgorithmTests,
                                   QasmReadoutNoiseTests,
                                   QasmResetNoiseTests,
                                   QasmPauliNoiseTests,
