+++ test/terra/backends/test_qasm_simulator_statevector.py
@@ -24,12 +24,15 @@ from test.terra.backends.qasm_simulator.qasm_measure import QasmMultiQubitMeasur
 from test.terra.backends.qasm_simulator.qasm_cliffords import QasmCliffordTests
 from test.terra.backends.qasm_simulator.qasm_cliffords import QasmCliffordTestsWaltzBasis
 from test.terra.backends.qasm_simulator.qasm_cliffords import QasmCliffordTestsMinimalBasis
-from test.terra.backends.qasm_simulator.qasm_noncliffords import QasmNonCliffordTests
+from test.terra.backends.qasm_simulator.qasm_noncliffords import QasmNonCliffordTestsTGate
+from test.terra.backends.qasm_simulator.qasm_noncliffords import QasmNonCliffordTestsCCXGate
+from test.terra.backends.qasm_simulator.qasm_noncliffords import QasmNonCliffordTestsCGates
 from test.terra.backends.qasm_simulator.qasm_noncliffords import QasmNonCliffordTestsWaltzBasis
 from test.terra.backends.qasm_simulator.qasm_noncliffords import QasmNonCliffordTestsMinimalBasis
 from test.terra.backends.qasm_simulator.qasm_unitary_gate import QasmUnitaryGateTests
 from test.terra.backends.qasm_simulator.qasm_unitary_gate import QasmDiagonalGateTests
 from test.terra.backends.qasm_simulator.qasm_initialize import QasmInitializeTests
+from test.terra.backends.qasm_simulator.qasm_multiplexer import QasmMultiplexerTests
 # Conditional instruction tests
 from test.terra.backends.qasm_simulator.qasm_conditional import QasmConditionalGateTests
 from test.terra.backends.qasm_simulator.qasm_conditional import QasmConditionalUnitaryTests
@@ -65,9 +68,10 @@ class StatevectorTests(
         QasmResetTests, QasmInitializeTests, QasmConditionalGateTests,
         QasmConditionalUnitaryTests, QasmConditionalKrausTests,
         QasmCliffordTests, QasmCliffordTestsWaltzBasis,
-        QasmCliffordTestsMinimalBasis, QasmNonCliffordTests,
+        QasmCliffordTestsMinimalBasis, QasmNonCliffordTestsTGate,
+        QasmNonCliffordTestsCCXGate, QasmNonCliffordTestsCGates,
         QasmNonCliffordTestsWaltzBasis, QasmNonCliffordTestsMinimalBasis,
-        QasmAlgorithmTests, QasmAlgorithmTestsWaltzBasis,
+        QasmMultiplexerTests, QasmAlgorithmTests, QasmAlgorithmTestsWaltzBasis,
         QasmAlgorithmTestsMinimalBasis, QasmUnitaryGateTests, QasmDiagonalGateTests,
         QasmReadoutNoiseTests, QasmPauliNoiseTests, QasmThreadManagementTests,
         QasmFusionTests, QasmDelayMeasureTests, QasmQubitsTruncateTests,
