+++ test/terra/reference/ref_2q_clifford.py
@@ -16,7 +16,6 @@ Test circuits and reference outputs for 2-qubit Clifford gate instructions.
 
 import numpy as np
 from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
-from test.terra.utils.multiplexer import multiplexer_multi_controlled_x
 
 
 # ==========================================================================
@@ -710,135 +709,3 @@ def swap_gate_unitary_nondeterministic():
                              [0, 0, 1, 1, 0, 0, 0, 0],
                              [0, 0, 1, -1, 0, 0, 0, 0]]) / np.sqrt(2))
     return targets
-
-def multiplexer_cx_gate_circuits_deterministic(final_measure=True):
-    """multiplexer-gate simulating cx gate, test circuits with deterministic counts."""
-    circuits = []
-    qr = QuantumRegister(2)
-    if final_measure:
-        cr = ClassicalRegister(2)
-        regs = (qr, cr)
-    else:
-        regs = (qr, )
-
-    num_control_qubits = 1
-
-    # CX01, |00> state
-    circuit = QuantumCircuit(*regs)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits), [qr[0], qr[1]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # CX10, |00> state
-    circuit = QuantumCircuit(*regs)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits), [qr[1], qr[0]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # CX01.(X^I), |10> state
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr[1])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits), [qr[0], qr[1]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # CX10.(I^X), |01> state
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr[0])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits), [qr[1], qr[0]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # CX01.(I^X), |11> state
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr[0])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits), [qr[0], qr[1]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # CX10.(X^I), |11> state
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr[1])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits), [qr[1], qr[0]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # CX01.(X^X), |01> state
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr)
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits), [qr[0], qr[1]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # CX10.(X^X), |10> state
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr)
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits), [qr[1], qr[0]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    return circuits
-
-def multiplexer_cx_gate_circuits_nondeterministic(final_measure=True):
-    """Multiplexer CX-like gate test circuits with non-deterministic counts."""
-    circuits = []
-    qr = QuantumRegister(2)
-    if final_measure:
-        cr = ClassicalRegister(2)
-        regs = (qr, cr)
-    else:
-        regs = (qr, )
-
-    # cx gate only has one control qubit
-    num_control_qubits = 1
-
-    # CX01.(I^H), Bell state
-    circuit = QuantumCircuit(*regs)
-    circuit.h(qr[0])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits), [qr[0], qr[1]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # CX10.(H^I), Bell state
-    circuit = QuantumCircuit(*regs)
-    circuit.h(qr[1])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits), [qr[1], qr[0]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-    return circuits
-
-def multiplexer_cx_gate_counts_deterministic(shots, hex_counts=True):
-    """ The counts are exactly the same as the cx gate """
-    return cx_gate_counts_deterministic(shots, hex_counts)
-
-def multiplexer_cx_gate_counts_nondeterministic(shots, hex_counts=True):
-    """ The counts are exactly the same as the cx gate """
-    return cx_gate_counts_nondeterministic(shots, hex_counts)
