+++ test/terra/reference/ref_non_clifford.py
@@ -15,7 +15,6 @@ Test circuits and reference outputs for non-Clifford gate instructions.
 
 import numpy as np
 from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
-from test.terra.utils.multiplexer import multiplexer_multi_controlled_x
 
 # ==========================================================================
 # T-gate
@@ -814,232 +813,6 @@ def ccx_gate_unitary_nondeterministic():
                   [0, 0, 0, 0, 1, 0, -1, 0], [0, 0, 0, 0, 0, 1, 0, -1]]) /
         np.sqrt(2))
     return targets
-
-
-# ==========================================================================
-# Multiplexer-gate (CCX-like)
-# ==========================================================================
-def multiplexer_ccx_gate_circuits_deterministic(final_measure=True):
-    """multiplexer-gate simulating ccx gate, test circuits with deterministic counts."""
-
-    circuits = []
-    qr = QuantumRegister(3)
-    if final_measure:
-        cr = ClassicalRegister(3)
-        regs = (qr, cr)
-    else:
-        regs = (qr, )
-
-    # because ccx has two control qubits and one target
-    num_control_qubits = 2
-
-    # CCX(0,1,2)
-    circuit = QuantumCircuit(*regs)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[0], qr[1], qr[2]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # (I^X^X).CCX(0,1,2).(I^X^X) -> |100>
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr[0])
-    circuit.barrier(qr)
-    circuit.x(qr[1])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[0], qr[1], qr[2]])
-    circuit.barrier(qr)
-    circuit.x(qr[0])
-    circuit.barrier(qr)
-    circuit.x(qr[1])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # (X^I^X).CCX(0,1,2).(X^I^X) -> |000>
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr[0])
-    circuit.barrier(qr)
-    circuit.x(qr[2])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[0], qr[1], qr[2]])
-    circuit.barrier(qr)
-    circuit.x(qr[0])
-    circuit.barrier(qr)
-    circuit.x(qr[2])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # (X^X^I).CCX(0,1,2).(X^X^I) -> |000>
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr[1])
-    circuit.barrier(qr)
-    circuit.x(qr[2])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[0], qr[1], qr[2]])
-    circuit.barrier(qr)
-    circuit.x(qr[1])
-    circuit.barrier(qr)
-    circuit.x(qr[2])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # CCX(2,1,0)
-    circuit = QuantumCircuit(*regs)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[2], qr[1], qr[0]])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # (I^X^X).CCX(2,1,0).(I^X^X) -> |000>
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr[0])
-    circuit.barrier(qr)
-    circuit.x(qr[1])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[2], qr[1], qr[0]])
-    circuit.barrier(qr)
-    circuit.x(qr[0])
-    circuit.barrier(qr)
-    circuit.x(qr[1])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # (X^I^X).CCX(2,1,0).(X^I^X) -> |000>
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr[0])
-    circuit.barrier(qr)
-    circuit.x(qr[2])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[2], qr[1], qr[0]])
-    circuit.barrier(qr)
-    circuit.x(qr[0])
-    circuit.barrier(qr)
-    circuit.x(qr[2])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # (X^X^I).CCX(2,1,0).(X^X^I) -> |001>
-    circuit = QuantumCircuit(*regs)
-    circuit.x(qr[1])
-    circuit.barrier(qr)
-    circuit.x(qr[2])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[2], qr[1], qr[0]])
-    circuit.barrier(qr)
-    circuit.x(qr[1])
-    circuit.barrier(qr)
-    circuit.x(qr[2])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    return circuits
-
-
-def multiplexer_ccx_gate_circuits_nondeterministic(final_measure=True):
-    """mukltiplexer CCX-like gate test circuits with non-deterministic counts."""
-    circuits = []
-    qr = QuantumRegister(3)
-    if final_measure:
-        cr = ClassicalRegister(3)
-        regs = (qr, cr)
-    else:
-        regs = (qr, )
-
-    # because ccx has two control qubits and one target
-    num_control_qubits = 2
-
-    # (I^X^I).CCX(0,1,2).(I^X^H) -> |000> + |101>
-    circuit = QuantumCircuit(*regs)
-    circuit.h(qr[0])
-    circuit.barrier(qr)
-    circuit.x(qr[1])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[0], qr[1], qr[2]])
-    circuit.barrier(qr)
-    circuit.x(qr[1])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # (I^I^X).CCX(0,1,2).(I^H^X) -> |000> + |110>
-    circuit = QuantumCircuit(*regs)
-    circuit.h(qr[1])
-    circuit.barrier(qr)
-    circuit.x(qr[0])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[0], qr[1], qr[2]])
-    circuit.barrier(qr)
-    circuit.x(qr[0])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # (I^X^I).CCX(2,1,0).(H^X^I) -> |000> + |101>
-    circuit = QuantumCircuit(*regs)
-    circuit.h(qr[2])
-    circuit.barrier(qr)
-    circuit.x(qr[1])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[2], qr[1], qr[0]])
-    circuit.barrier(qr)
-    circuit.x(qr[1])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    # (X^I^I).CCX(2,1,0).(X^H^I) -> |000> + |011>
-    circuit = QuantumCircuit(*regs)
-    circuit.h(qr[1])
-    circuit.barrier(qr)
-    circuit.x(qr[2])
-    circuit.barrier(qr)
-    circuit.append(multiplexer_multi_controlled_x(num_control_qubits),
-                   [qr[2], qr[1], qr[0]])
-    circuit.barrier(qr)
-    circuit.x(qr[2])
-    if final_measure:
-        circuit.barrier(qr)
-        circuit.measure(qr, cr)
-    circuits.append(circuit)
-
-    return circuits
-
-
-def multiplexer_ccx_gate_counts_deterministic(shots, hex_counts=True):
-    """ The counts are exactly the same as the ccx gate """
-    return ccx_gate_counts_deterministic(shots, hex_counts)
-
-
-def multiplexer_ccx_gate_counts_nondeterministic(shots, hex_counts=True):
-    """ The counts are exactly the same as the ccx gate """
-    return ccx_gate_counts_nondeterministic(shots, hex_counts)
 
 
 # ==========================================================================
